Refactors/adds feature to Sounds
Sounds will now load numbered sounds.
A sound played by a string "cue" will do the following:
1. If the sounds for that "cue" have been loaded (or assigned via the
inspector), then it selects one of the loaded sounds, and plays it.
2. If the sounds have not been loaded yet- it attempts to load the sound
settings.
If it finds a GameObject named "cue" with an AudioSource in the
resources folder, it will assign that object to the resulting settings
If it finds an AudioClip in the Resources folder called "cue", it will
assign that sound to the resulting settings.
If it does not, it will look for sounds named "cue0" and "cue1". if it
finds either of these, it will continue searching from "cue2" until it
stops finding sounds. All of those sounds are then added into the
resulting Sound cue.
if no sounds are found, an empty cue is added, and no sound is played.