Minor changes
FancyText velocity now scales with screensize.

HoverEntropy also changes the position in time of the Hover object's
oscilators.

Adds scale property to Screen, which is (height/720). this can be used
to make sure that all GUI elements have a consistent scale

For some reason, a few line endings changed in GUIUtils