Much changes Adds Update() and LateUpdate() functionality to windows registered to GUIRoot Adds virtual Update and LateUpdate functions to ZWindow Adds Focus() and Unfocus() member functions to ZWindow Adds an Anchors static class for holding common Rects for UGUI anchors Adds NetExec shell to DevConsole Changes ReflectionUtils to use += instead of Append for StringBuilder usage. Adds a list of assemblies to ReflectionUtils Adds AttributeSummary to ReflectionUtils
New classes:
UGUIzTextField - Send input from a user to some location via reflection
TextMeshReflectionFill - take information from some location via reflection and fill a Text object with that information
TextRect - sets the RectTransform size based on the measurements of text on the object.
TextLog - A simple List<string> that holds a number of strings, up to a max. kicks out the oldest string when a new one is added at the max size.