Changes to GUIRoot/DevConsole and various extensions.
Adds Minimal<T>() and Maximal<T>() extensions on IEnumerable<T>, taking a Func<T, (Numeric)> to select a T, and returns the smallest or largest T in a collection.

Adds UnixTimestamp() to DateTime, returning a long representing the DateTime as a UnixTimestamp.

Modifies SplitUnlessInContainer and DevConsole to fix issue with multiple +/- aliases on a single key not always firing - commands.

For GUIRoot, there are a few changes:
Adds Hold binds and registration by name to GUIRoot, allowing windows to be kept open while a key is held down, and windows to be registered simply by name, without being bound.

Remove now unbinds/unregisters any windows with a given bind/name.
Unbind simply removes any windows bound to press/hold for a given key bind.
Bind() binds a window to toggle on key-press.
BindHold() binds a window to stay open while a key is held down.
Register() adds a window by name, but doesn't change its status when a key is pressed.